requirements.txt
altgraph==0.10.2
appdirs==1.4.3
appnope==0.1.0
Babel==2.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
beautifulsoup4==4.5.3
bonjour-py==0.3
brukva==0.0.1
certifi==2017.1.23
chardet==3.0.2
daemon==1.0
decorator==4.0.9
Django==1.10.dev20160317162119
django-annoying==0.10.1
django-bootstrap3==8.2.2
django-ckeditor==5.2.2
django-extensions==1.7.8
django-haystack==2.5.0
django-machina==0.5.3
django-markdown==0.8.4
django-material==0.8.0
django-materialize-form==0.1.1
django-mptt==0.8.5
django-phonenumber-field==1.1.0
django-socketio==0.3.9
django-webtest==1.9.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
gevent-socketio==0.2.1
gevent-websocket==0.10.1
Gnotty==0.2.7
gnureadline==6.3.3
httpie==0.9.9
ipython==4.2.0
ipython-genutils==0.1.0
ldif3==3.1.1
macholib==1.5.1
Markdown==2.6.6
markdown2==2.3.4
matplotlib==1.3.1
modulegraph==0.10.4
mysql-connector-python-cext==2.1.3
MySQL-python==1.2.5
numpy==1.8.0rc1
oauthlib==2.0.2
packaging==16.8
pathlib2==2.1.0
pexpect==4.1.0
phonenumberslite==7.5.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
py2app==0.7.3
pybbm==0.17.3
Pygments==2.2.0
PyJWT==1.4.2
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-card-me==0.9.3
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2013.7
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
scipy==0.13.0b1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.4.1
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sorl-thumbnail==12.4a1
sphinx-me==0.3
surlex==0.2.0
tornado==4.4.3
traitlets==4.2.1
unicodecsv==0.14.1
Unidecode==0.4.19
vboxapi==1.0
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
xattr==0.6.4
xmltodict==0.10.2
zope.interface==4.1.1
